pyproject.toml
[project]
name = "whisper-transcriber"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "openai-whisper>=20240930",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/whisper_transcriber"]
